+++ app/src/main/java/com/jacpalberto/devcomms/Agenda/AgendaByDateFragment.kt
@@ -11,8 +11,9 @@ import android.view.View
 import android.view.ViewGroup
 import com.jacpalberto.devcomms.R
 import com.jacpalberto.devcomms.adapters.PagerAdapter
+import com.jacpalberto.devcomms.data.DataResponse
 import com.jacpalberto.devcomms.data.DevCommsEvent
-import com.jacpalberto.devcomms.data.DevCommsEventList
+import com.jacpalberto.devcomms.data.EventListWrapper
 import com.jacpalberto.devcomms.events.EventsViewModel
 import com.jacpalberto.devcomms.utils.doOnTabSelected
 import kotlinx.android.synthetic.main.fragment_events_by_date.*
@@ -46,9 +47,9 @@ class AgendaByDateFragment : Fragment() {
         viewModel?.fetchFavoriteEvents()?.observe(this, Observer { filterEventsByDate(it) })
     }
 
-    private fun filterEventsByDate(eventList: DevCommsEventList?) {
+    private fun filterEventsByDate(eventList: DataResponse<List<DevCommsEvent>>?) {
         eventList?.let { events ->
-            val eventsMap = events.eventList.groupBy { it.startDateString }
+            val eventsMap = events.data.groupBy { it.startDateString }
             setupTabLayout(eventsMap.keys)
             setupViewPager(eventsMap)
         }
@@ -56,9 +57,7 @@ class AgendaByDateFragment : Fragment() {
 
     private fun setupViewPager(eventsMap: Map<String?, List<DevCommsEvent>>) {
         val fragmentList = mutableListOf<Fragment>()
-        eventsMap.forEach { _, list ->
-            fragmentList.add(AgendaFragment.newInstance(DevCommsEventList(list)))
-        }
+        eventsMap.forEach { _, list -> fragmentList.add(AgendaFragment.newInstance(EventListWrapper(list))) }
         viewPager.adapter = activity?.supportFragmentManager?.let { PagerAdapter(childFragmentManager, fragmentList) }
     }