+++ app/src/main/java/com/jacpalberto/devcomms/sponsorDetail/SponsorDetailActivity.kt
@@ -3,9 +3,9 @@ package com.jacpalberto.devcomms.sponsorDetail
 import android.content.Context
 import android.content.Intent
 import android.os.Build
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.support.v4.content.ContextCompat
+import android.support.v7.app.AppCompatActivity
 import android.text.Spannable
 import android.text.SpannableString
 import android.text.TextPaint