+++ app/src/main/java/com/jacpalberto/devcomms/sponsors/SponsorsModel.kt
@@ -1,11 +1,9 @@
 package com.jacpalberto.devcomms.sponsors
 
 import android.os.AsyncTask
-import android.util.Log
 import com.jacpalberto.devcomms.DevCommsApp
-import com.jacpalberto.devcomms.data.DataState
+import com.jacpalberto.devcomms.data.DataResponse
 import com.jacpalberto.devcomms.data.Sponsor
-import com.jacpalberto.devcomms.data.SponsorList
 
 
 /**
@@ -14,25 +12,23 @@ import com.jacpalberto.devcomms.data.SponsorList
 class SponsorsModel {
     private val db by lazy { DevCommsApp.database }
     private val sponsorsDao by lazy { db!!.sponsorsDao() }
-    private var sponsorsList: List<Sponsor> = emptyList()
+    private var sponsors: List<Sponsor> = emptyList()
     private val repository = SponsorRepository()
 
-    fun fetchSponsors(onResult: (SponsorList) -> Unit) {
-        sponsorsList = emptyList()
-        var sponsorResult: SponsorList
+    fun fetchSponsors(onResult: (response: DataResponse<List<Sponsor>>) -> Unit) {
+        sponsors = emptyList()
+        val finalResponse = DataResponse<List<Sponsor>>(emptyList())
+
         repository.fetchSponsors { response ->
-            if (response.status == DataState.FAILURE || response.status == DataState.ERROR) {
-                sponsorsList = sponsorsDao.getList()
-                sponsorResult = if (sponsorsList.isEmpty())
-                    SponsorList(sponsorsList, status = DataState.FAILURE)
-                else SponsorList(sponsorsList, status = DataState.SUCCESS)
-                onResult(sponsorResult)
+            if (response.isStatusFailedOrError()) {
+                sponsors = sponsorsDao.getList()
+                if (sponsors.isEmpty()) onResult(finalResponse.apply { setFailureStatus() })
+                else onResult(finalResponse.apply { updateSuccessValue(sponsors) })
             } else {
-                saveAll(response.sponsorList)
+                saveAll(response.data)
                 AsyncTask.execute {
-                    sponsorsList = sponsorsDao.getList()
-                    sponsorResult = SponsorList(sponsorsList, status = DataState.SUCCESS)
-                    onResult(sponsorResult)
+                    sponsors = sponsorsDao.getList()
+                    onResult(finalResponse.apply { updateSuccessValue(sponsors) })
                 }
             }
         }