+++ app/src/main/java/com/jacpalberto/devcomms/sponsors/SponsorsViewModel.kt
@@ -2,17 +2,17 @@ package com.jacpalberto.devcomms.sponsors
 
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
-import android.util.Log
-import com.jacpalberto.devcomms.data.SponsorList
+import com.jacpalberto.devcomms.data.DataResponse
+import com.jacpalberto.devcomms.data.Sponsor
 
 /**
  * Created by Alberto Carrillo on 8/13/18.
  */
 class SponsorsViewModel : ViewModel() {
-    private var sponsors: MutableLiveData<SponsorList>? = null
+    private var sponsors: MutableLiveData<DataResponse<List<Sponsor>>>? = null
     private val model = SponsorsModel()
 
-    fun fetchSponsors(): MutableLiveData<SponsorList>? {
+    fun fetchSponsors(): MutableLiveData<DataResponse<List<Sponsor>>>? {
         if (sponsors == null) {
             sponsors = MutableLiveData()
             model.fetchSponsors { sponsors?.postValue(it) }