+++ app/src/main/java/com/jacpalberto/devcomms/utils/extensions.kt
@@ -15,7 +15,6 @@ import android.text.SpannableString
 import android.text.TextPaint
 import android.text.method.LinkMovementMethod
 import android.text.style.ClickableSpan
-import android.util.Log
 import android.view.View
 import android.widget.TextView
 import android.widget.Toast
@@ -100,7 +99,6 @@ fun createAnimationOnAnimationEnd(animator: (p0: Animator?) -> Unit = {}): Anima
         }
 
         override fun onAnimationEnd(animation: Animator?) {
-            Log.d("AnimationEnd", "Ending now")
             animator(animation)
         }