+++ CHANGELOG.md
@@ -13,10 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Version to the SplashScreen based on `BuildConfig.VERSION_NAME`
 
 ### Changed
-
 - Updated version on dependecy: `com.google.firebase:firebase-database` 16.0.1 -> 16.0.2
 - Updated version on dependecy: `com.google.firebase:firebase-database` 17.3.1 -> 17.3.2
-
+- Updated responsesModels to DataResponse to handle errors and status
 
 [Unreleased]: https://github.com/GDLDevComms/DevComms/compare/v1.0.1...HEAD
 [1.0.1]: https://github.com/GDLDevComms/DevComms/compare/v1.0.1...v1.0.0